--- APIMobile/App.xaml.cs
@@ -1,14 +1,0 @@
-﻿namespace APIMobile;
-
-public partial class App : Application
-{
-	public App()
-	{
-		InitializeComponent();
-	}
-
-	protected override Window CreateWindow(IActivationState? activationState)
-	{
-		return new Window(new AppShell());
-	}
-}
