--- APIMobile/MauiProgram.cs
@@ -1,24 +1,0 @@
-﻿using Microsoft.Extensions.Logging;
-
-namespace APIMobile;
-
-public static class MauiProgram
-{
-	public static MauiApp CreateMauiApp()
-	{
-		var builder = MauiApp.CreateBuilder();
-		builder
-			.UseMauiApp<App>()
-			.ConfigureFonts(fonts =>
-			{
-				fonts.AddFont("OpenSans-Regular.ttf", "OpenSansRegular");
-				fonts.AddFont("OpenSans-Semibold.ttf", "OpenSansSemibold");
-			});
-
-#if DEBUG
-		builder.Logging.AddDebug();
-#endif
-
-		return builder.Build();
-	}
-}
