+++ MauiApp1/App.xaml.cs
@@ -1,0 +1,15 @@
+﻿namespace MauiApp1
+{
+    public partial class App : Application
+    {
+        public App()
+        {
+            InitializeComponent();
+        }
+
+        protected override Window CreateWindow(IActivationState? activationState)
+        {
+            return new Window(new AppShell());
+        }
+    }
+}
