+++ MauiApp1/AppShell.xaml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <Shell
-    x:Class="APIMobile.AppShell"
+    x:Class="MauiApp1.AppShell"
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    xmlns:local="clr-namespace:APIMobile"
-    Title="APIMobile">
+    xmlns:local="clr-namespace:MauiApp1"
+    Title="MauiApp1">
 
     <ShellContent
         Title="Home"
