+++ MauiApp1/AppShell.xaml.cs
@@ -1,0 +1,10 @@
+﻿namespace MauiApp1
+{
+    public partial class AppShell : Shell
+    {
+        public AppShell()
+        {
+            InitializeComponent();
+        }
+    }
+}
