+++ MauiApp1/MauiApp1.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.13.35931.197 d17.13
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MauiApp1", "MauiApp1.csproj", "{B837EB2F-5F55-42FE-B1F2-1E21492B1ED7}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{B837EB2F-5F55-42FE-B1F2-1E21492B1ED7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B837EB2F-5F55-42FE-B1F2-1E21492B1ED7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B837EB2F-5F55-42FE-B1F2-1E21492B1ED7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B837EB2F-5F55-42FE-B1F2-1E21492B1ED7}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {F5AD1A5C-019B-4025-A39C-E0701E562168}
+	EndGlobalSection
+EndGlobal
