+++ MauiApp1/Platforms/Android/AndroidManifest.xml
