+++ MauiApp1/Platforms/MacCatalyst/AppDelegate.cs
@@ -1,0 +1,10 @@
+﻿using Foundation;
+
+namespace MauiApp1
+{
+    [Register("AppDelegate")]
+    public class AppDelegate : MauiUIApplicationDelegate
+    {
+        protected override MauiApp CreateMauiApp() => MauiProgram.CreateMauiApp();
+    }
+}
