+++ MauiApp1/Platforms/MacCatalyst/Entitlements.plist
