+++ MauiApp1/Platforms/iOS/Resources/PrivacyInfo.xcprivacy
