+++ MauiApp1/Properties/launchSettings.json
