+++ MauiApp1/Resources/Images/dotnet_bot.png
