+++ MauiApp1/Resources/Raw/AboutAssets.txt
